shard/requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
aniso8601==7.0.0
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
discord>=1.0.1
discord.py>=1.3.0
emoji==0.5.1
entrypoints==0.3
flake8==3.7.7
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.0.1
logzero==1.5.0
lxml==4.2.5
lyricwikia==0.1.9
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
multidict==4.5.2
mutagen==1.41.1
numpy==1.17.0
pafy==0.5.4
pathlib==1.0.1
Pillow==6.2.1
psycopg2-binary==2.7.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyNaCl==1.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
python-twitter==3.5
pytz==2018.7
requests==2.21.0
setuptools==39.0.1
six==1.12.0
spotipy==2.4.4
SQLAlchemy==1.3.0
titlecase==0.12.0
typing-extensions==3.6.6
Unidecode==1.0.23
urllib3==1.24.2
uWSGI==2.0.18
websockets==6.0
Werkzeug==0.15.4
yarl==1.3.0
youtube-dl
pyyaml
pika
aio-pika
